requirements.txt
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.3
blessings==1.7
bs4==0.0.2
cachetools==5.3.3
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
colour-runner==0.1.1
comm==0.1.3
contourpy==1.1.0
coverage==7.6.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
deepdiff==7.0.1
distlib==0.3.8
executing==1.2.0
filelock==3.15.4
fonttools==4.41.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.25.0
ipython==8.12.2
jedi==0.19.0
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
matplotlib==3.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.24.4
ordered-set==4.1.0
packaging==24.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==4.2.2
pluggy==1.5.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
pyproject-api==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
requests==2.31.0
rootpath==0.1.1
six==1.16.0
soupsieve==2.5
stack-data==0.6.2
termcolor==2.4.0
tomli==2.0.1
tornado==6.3.2
tox==4.16.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.26.3
wcwidth==0.2.6
wget==3.2
zipp==3.16.2
